drive setup

+++ src/data/earthengine/era5.py
@@ -34,8 +34,7 @@ def get_single_era5_image(region: ee.Geometry, start_date: date, end_date: date)
     ).first()
 
     if image.getInfo() is None:
-        return create_placeholder(region, ERA5_BANDS)
+        return create_placeholder(region, ERA5_BANDS).toDouble()
 
-    # has to be double to be compatible with the sentinel 1 imagery, which is in
-    # float64
+    # all imagery has to have the same data type to be compatible
     return image.toDouble()
