+++ src/data/earthengine/viirs.py
@@ -34,7 +34,7 @@ def get_single_viirs_500m_image(region: ee.Geometry, start_date: date, end_date:
     ).first()
 
     if image.getInfo() is None:
-        return create_placeholder(region, VIIRS_BANDS_500m)
+        return create_placeholder(region, VIIRS_BANDS_500m).toDouble()
 
     # has to be double to be compatible with the sentinel 1 imagery, which is in
     # float64
@@ -58,8 +58,7 @@ def get_single_viirs_1000m_image(region: ee.Geometry, start_date: date, end_date
     ).first()
 
     if image.getInfo() is None:
-        return create_placeholder(region, VIIRS_BANDS_1000m)
+        return create_placeholder(region, VIIRS_BANDS_1000m).toDouble()
 
-    # has to be double to be compatible with the sentinel 1 imagery, which is in
-    # float64
+    # all imagery has to have the same data type to be compatible
     return image.toDouble()
